feat: protect journal retention for projections

+++ hat/hatCache/journal_segments.go
@@ -261,6 +261,9 @@ func (journal *CommandJournal) pruneSegmentsLocked() error {
 		if journal.outboxRetainFrom > 0 && segment.end >= journal.outboxRetainFrom {
 			break
 		}
+		if projectionThrough, protected := journal.projectionRetentionThroughLocked(); protected && segment.end > projectionThrough {
+			break
+		}
 		if err := os.Remove(segment.path); err != nil {
 			return err
 		}
